requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11.18
django-forms-bootstrap==3.1.0
django-geoposition==0.3.1
django-storages==1.7.1
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
stripe==2.21.0
urllib3==1.24.1
